+++ spec/components/email_spec.rb
@@ -107,6 +107,14 @@ describe Minfraud::Components::Email do
           input:  'Test+alias@maxmind.com',
           output: '977577b140bfb7c516e4746204fbdb01',
         },
+        {
+          input:  "bu\u0308cher@example.com",
+          output: '53550c712b146287a2d0dd30e5ed6f4b',
+        },
+        {
+          input:  "b\u00FCcher@example.com",
+          output: '53550c712b146287a2d0dd30e5ed6f4b',
+        },
       ]
 
       tests.each do |i|