+++ CHANGELOG.md
@@ -26,6 +26,8 @@
   used. For example, `gmali.com` will become `gmail.com`.
 * When `hash_address` is used, all trailing periods are now removed from an
   email address domain. Previously only a single period was removed.
+* When `hash_address` is used, the local part of an email address is now
+  normalized to NFC.
 
 ## v2.4.0 (2024-01-12)